% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aviationgeoms-package.R
\docType{package}
\name{aviationgeoms}
\alias{aviationgeoms}
\alias{aviationgeoms-package}
\title{ggplot2 for aviation}
\description{
This package provides some simple function calls to 
add aviation layers to \code{ggplot2} figures.
}
