% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_procedure.R, R/stat_procedure.R
\name{geom_procedure}
\alias{geom_procedure}
\alias{stat_procedure}
\title{Display lines representing each procedure}
\usage{
geom_procedure(mapping = NULL, data = NULL, position = "identity",
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE, ...)

stat_procedure(mapping = NULL, data = NULL, geom = "path",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{logical. Should this layer remove NA values}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{...}{Additional parameters to the \code{geom} and \code{stat}.}

\item{geom}{The geometric object to use display the data}

\item{show.legnd}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}
}
\description{
Given a data.frame containing a list of SIDs, STARs, Approaches, and Enroute procedures, create a plot.  Currently only IF, TF, and RF legs are supported.
HM legs are removed.  If there are still unsupported leg types, the first continous subset of known leg types is plotted.

If plotting more than one procedure you will generally want to specify the group aesthetic as \code{group=paste(procedure, transistion)}

\code{stat_procedure} and \code{geom_procedure} understand the following aesthetics. Required aesthetics are in bold and if RF legs are present the italicized aesthetics are also required.
\itemize{
\item \bold{x}
\item \bold{y}
\item \bold{leg_type}
\item \emph{rf_center_longitude}
\item \emph{rf_center_latitude}
\item \emph{turn_direction}
\item alpha
\item colour
\item fill
\item linetype
\item size
}
}
\examples{
\dontrun(
# given some procedure data with the latitude, longitude, procedure/transition names, and ARINC424 leg types
library(ggplot2)
ggplot(procs, aes(y=latitude, x=longitude, group=paste(procedure, transition), leg_type=leg_type)) + geom_procedure()
)
}
