% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_rangering.R, R/stat_rangering.R
\name{geom_rangering}
\alias{geom_rangering}
\alias{stat_rangering}
\title{plot a range ring}
\usage{
geom_rangering(mapping = NULL, data = NULL, position = "identity",
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE, ...)

stat_rangering(mapping = NULL, data = NULL, geom = "polygon",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{logical. Should this layer remove NA values}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{...}{Additional parameters to the \code{geom} and \code{stat}.}

\item{geom}{The geometric object to use display the data}

\item{show.legnd}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}
}
\description{
Range rings provide some perspective in plan view plots. This function
makes drawing a circle using sperical coordinates on a map simple.  \code{geom_rangering} is backed by \code{geom_polygon}
so it is possible to plot "donuts" by specifying and \code{inner_radius} (you will probably also want to specify \code{fill})

\code{stat_rangering} and\code{stat_rangering} understand the following aesthetics (required aesthetics are in bold)
\itemize{
\item \bold{x}
\item \bold{y}
\item \bold{radius}
\item inner_radius
\item alpha
\item colour
\item fill
\item linetype
\item size
}
}
\examples{
# plot range ring about Denver
den_rr <- data.frame(x=-104.6732, y= 39.86167, radius=40)
ggplot(den_rr, aes(x=x,y=y, radius=radius)) + geom_rangering() + coord_map()

# plot range donut about Denver
ggplot(den_rr, aes(x=x,y=y, radius=radius)) + geom_rangering(inner_radius=30, fill="blue") + coord_map()

}
