% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_runway.R, R/stat_runway.R
\name{geom_runway}
\alias{geom_runway}
\alias{stat_runway}
\title{plot runways}
\usage{
geom_runway(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

stat_runway(
  mapping = NULL,
  data = NULL,
  geom = "polygon",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{logical. Should this layer remove NA values}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{...}{Additional parameters to the \code{geom} and \code{stat}.}

\item{geom}{The geometric object to use display the data}

\item{show.legnd}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}
}
\description{
Given runway names and end points lay down a polygon plot

\code{stat_runway} and\code{stat_runway} understand the following aesthetics (required aesthetics are in bold).  The first five
correspond to the endpoints of the runway and it's width in feet.  these names match the column names of the data.frame returned
by \code{caasddb::fetch_runways()}
\itemize{
\item \bold{latitude_AER}
\item \bold{longitude_AER}
\item \bold{latitude_DER}
\item \bold{longitude_DER}
\item width, defaults to runway width of 150ft if not provided
\item alpha
\item colour
\item fill
\item linetype
\item size
}
}
\examples{
library(tibble)
library(ggplot2)

# define data with the latitude/longitude for the arrival/departure ends of the runway (AER/DER)
# DEN example derived from https://www.airnav.com/airport/KDEN
den_rwys <- tribble(
  ~runway, ~latitude_AER, ~longitude_AER, ~latitude_DER, ~longitude_DER,
  "7/25", 39+50.457/60, -1*(104+43.599/60), 39+50.439/60, -1*(104+41.036/60),
  "8/26", 39+52.653/60, -1*(104+39.734/60), 39+52.635/60, -1*(104+37.169/60),
  "16L/34R", 39+53.822/60, -1*(104+41.208/60), 39+51.846/60, -1*(104+41.231/60),
  "16R/34L", 39+53.748/60, -1*(104+41.765/60), 39+51.113/60, -1*(104+41.795/60),
  "17L/35R", 39+51.897/60, -1*(104+38.478/60), 39+49.921/60, -1*(104+38.503/60),
  "17R/35L", 39+51.675/60, -1*(104+39.609/60), 39+49.699/60, -1*(104+39.633/60)
)

ggplot(den_rwys, 
       aes(latitude_AER=latitude_AER, longitude_AER=longitude_AER,
           latitude_DER=latitude_DER, longitude_DER=longitude_DER, 
           group=runway, fill=runway)) + 
  geom_runway()

}
